reorganized classes to reuse FileSystem in Pirum
