+++ classes/builders/Builder.php
@@ -1,0 +1,8 @@
+<?php
+
+interface Builder
+{
+	public function build(FileSystem $fs);
+}
+
+?>
