+++ classes/builders/PearPackage_Builder.php
@@ -1,12 +1,5 @@
 <?php
-/**
- * Created by JetBrains PhpStorm.
- * User: fqqdk
- * Date: 7/4/11
- * Time: 3:37 PM
- * To change this template use File | Settings | File Templates.
- */
- 
+
 class PearPackage_Builder implements Builder
 {
 	public function  __construct($targetDir)
@@ -22,3 +15,5 @@ class PearPackage_Builder implements Builder
 		chdir($oldCwd);
     }
 }
+
+?>
